Implement logic for checking favorite flights in FavoriteViewModel
This commit introduces a new function `isFavorite` to the `FavoriteViewModel`. This function checks if a flight, identified by its departure and destination codes, is present in the list of favorite flights.

The `setFavoriteFligth` function's parameter type has also been updated from `Airport` to `Favorite` to align with the data model.