Display list of favorite flights
The FavoriteScreen now displays a list of favorite flights using a LazyColumn.
Each item in the list is represented by an AirportItem composable.
If there are no favorite flights, a message indicating this is displayed.
The FavoriteViewModel is no longer directly accessed in FavoriteScreen; instead, the UI state is passed as a parameter.